+++ components/transaction/TransactionDetails.tsx
@@ -170,11 +170,11 @@ const TransactionDetails = ({ hash }: TxDetailsProps) => {
             <Typography variant="body2" component="h2">
               Hash
             </Typography>
-            <Typography variant="body2" component="h2">
-              {data.transaction && data.transaction.hash
-                ? data.transaction.hash
-                : <NotAvailable variant="body2" />}
-            </Typography>
+            {data.transaction && data.transaction.hash ?
+              <Typography variant="body2" component="h2">
+                {data.transaction.hash}
+              </Typography>
+              : <NotAvailable variant="body2" />}
             <Divider variant="middle" className={classes.divider} />
           </Grid>
 
@@ -217,7 +217,7 @@ const TransactionDetails = ({ hash }: TxDetailsProps) => {
               Status
             </Typography>
             <Typography variant="body2" component="h2">
-              {data.transaction && !data.transaction.pending ? (
+              {data.transaction && data.transaction.pending ? (
                 <Chips actionResult="Pending" />
               ) : (
                   <Chips actionResult="Success" />
@@ -231,8 +231,8 @@ const TransactionDetails = ({ hash }: TxDetailsProps) => {
               From
             </Typography>
 
-            <Typography variant="body2" component="h2">
-              {data.transaction && data.transaction.from ? (
+            {data.transaction && data.transaction.from ?
+              <Typography variant="body2" component="h2">
                 <Link
                   href="/account/[account]/"
                   as={`/account/${data.transaction.from.address}`}
@@ -240,8 +240,8 @@ const TransactionDetails = ({ hash }: TxDetailsProps) => {
                 >
                   {data.transaction.from.address}
                 </Link>
-              ) : <NotAvailable variant="body2" />}
             </Typography>
+              : <NotAvailable variant="body2" />}
 
             <Divider variant="middle" className={classes.divider} />
           </Grid>
@@ -250,8 +250,8 @@ const TransactionDetails = ({ hash }: TxDetailsProps) => {
             <Typography variant="body2" component="h2">
               To
             </Typography>
-            <Typography variant="body2" component="h2">
-              {data.transaction && data.transaction.to ? (
+            {data.transaction && data.transaction.to ?
+              <Typography variant="body2" component="h2">
                 <Link
                   href="/account/[account]/"
                   as={`/account/${data.transaction.to.address}`}
@@ -259,8 +259,9 @@ const TransactionDetails = ({ hash }: TxDetailsProps) => {
                 >
                   {data.transaction.to.address}
                 </Link>
-              ) : <NotAvailable variant="body2" />}
-            </Typography>
+              </Typography>
+              : <NotAvailable variant="body2" />}
+
             <Divider variant="middle" className={classes.divider} />
           </Grid>
 
@@ -268,11 +269,12 @@ const TransactionDetails = ({ hash }: TxDetailsProps) => {
             <Typography variant="body2" component="h2">
               Value
             </Typography>
-            <Typography variant="body2" component="h2">
-              {data.transaction && data.transaction.value
-                ? data.transaction.value
-                : <NotAvailable variant="body2" />}
-            </Typography>
+
+            {data.transaction && data.transaction.value ?
+              <Typography variant="body2" component="h2">
+                {data.transaction.value}
+              </Typography>
+              : <NotAvailable variant="body2" />}
             <Divider variant="middle" className={classes.divider} />
           </Grid>
 
@@ -280,16 +282,15 @@ const TransactionDetails = ({ hash }: TxDetailsProps) => {
             <Typography variant="body2" component="h2">
               Block Height
             </Typography>
-            <Typography variant="body2" component="h2">
-              <Link
-                href={`/block/${data.transaction.blockNumber}`}
-                color="secondary"
-              >
-                {data.transaction && data.transaction.blockNumber
-                  ? data.transaction.blockNumber
-                  : <NotAvailable variant="body2" />}
-              </Link>
-            </Typography>
+            {data.transaction && data.transaction.blockNumber ?
+              < Typography variant="body2" component="h2">
+                <Link
+                  href={`/block/${data.transaction.blockNumber}`}
+                  color="secondary"
+                > {data.transaction.blockNumber}
+                </Link>
+              </Typography>
+              : <NotAvailable variant="body2" />}
             <Divider variant="middle" className={classes.divider} />
           </Grid>
 
@@ -297,11 +298,11 @@ const TransactionDetails = ({ hash }: TxDetailsProps) => {
             <Typography variant="body2" component="h2">
               Nonce
             </Typography>
-            <Typography variant="body2" component="h2">
-              {data.transaction && data.transaction.nonce
-                ? data.transaction.nonce
-                : <NotAvailable variant="body2" />}
-            </Typography>
+            {data.transaction && data.transaction.nonce ?
+              < Typography variant="body2" component="h2">
+                {data.transaction.nonce}
+              </Typography>
+              : <NotAvailable variant="body2" />}
             <Divider variant="middle" className={classes.divider} />
           </Grid>
 
@@ -309,11 +310,11 @@ const TransactionDetails = ({ hash }: TxDetailsProps) => {
             <Typography variant="body2" component="h2">
               Transaction Fee
             </Typography>
-            <Typography variant="body2" component="h2">
-              {data.transaction && data.transaction.feeCurrency
-                ? data.transaction.feeCurrency + "cGLD"
-                : <NotAvailable variant="body2" />}
+            {data.transaction && data.transaction.feeCurrency ?
+              <Typography variant="body2" component="h2">
+                {data.transaction.feeCurrency} + "cGLD"
             </Typography>
+              : <NotAvailable variant="body2" />}
             <Divider variant="middle" className={classes.divider} />
           </Grid>
 
@@ -321,11 +322,11 @@ const TransactionDetails = ({ hash }: TxDetailsProps) => {
             <Typography variant="body2" component="h2">
               Fee Receipient
             </Typography>
-            <Typography variant="body2" component="h2">
-              {data.transaction && data.transaction.gatewayFeeRecipient
-                ? data.transaction.gatewayFeeRecipient
-                : <NotAvailable variant="body2" />}
-            </Typography>
+            {data.transaction && data.transaction.gatewayFeeRecipient
+              ? <Typography variant="body2" component="h2">
+                {data.transaction.gatewayFeeRecipient}
+              </Typography>
+              : <NotAvailable variant="body2" />}
             <Divider variant="middle" className={classes.divider} />
           </Grid>
 
@@ -333,11 +334,13 @@ const TransactionDetails = ({ hash }: TxDetailsProps) => {
             <Typography variant="body2" component="h2">
               Gate Fee
             </Typography>
-            <Typography variant="body2" component="h2">
-              {data.transaction && data.transaction.gatewayFee
-                ? data.transaction.gatewayFee
-                : <NotAvailable variant="body2" />}
-            </Typography>
+            {data.transaction && data.transaction.gatewayFee
+              ?
+              <Typography variant="body2" component="h2">
+                {data.transaction.gatewayFee}
+              </Typography>
+              : <NotAvailable variant="body2" />}
+
             <Divider variant="middle" className={classes.divider} />
           </Grid>
 
@@ -345,9 +348,11 @@ const TransactionDetails = ({ hash }: TxDetailsProps) => {
             <Typography variant="body2" component="h2">
               Transaction Speed
             </Typography>
-            <Typography variant="body2" component="h2">
-              1.5 seconds
-            </Typography>
+            {data.transaction && data.transaction.speed ?
+              <Typography variant="body2" component="h2">
+                {data.transaction.speed}
+              </Typography>
+              : <NotAvailable variant="body2" />}
             <Divider variant="middle" className={classes.divider} />
           </Grid>
 
@@ -406,11 +411,11 @@ const TransactionDetails = ({ hash }: TxDetailsProps) => {
             <Typography variant="body2" component="h2">
               Gas Used
             </Typography>
-            <Typography variant="body2" component="h2">
-              {data.transaction && data.transaction.gas
-                ? data.transaction.gas
-                : <NotAvailable variant="body2" />}
-            </Typography>
+            {data.transaction && data.transaction.gas ?
+              <Typography variant="body2" component="h2">
+                {data.transaction.gas}
+              </Typography>
+              : <NotAvailable variant="body2" />}
             <Divider variant="middle" className={classes.divider} />
           </Grid>
 
@@ -418,9 +423,11 @@ const TransactionDetails = ({ hash }: TxDetailsProps) => {
             <Typography variant="body2" >
               Gas Limit
             </Typography>
-            <Typography variant="body2">
-              20,000.000
-            </Typography>
+            {data.transaction && data.transaction.gasLimit
+              ? <Typography variant="body2">
+                {data.transaction.gasLimit}
+              </Typography>
+              : <NotAvailable variant="body2" />}
           </Grid>
         </Grid>
       </CardContent>
@@ -435,7 +442,7 @@ const TransactionDetails = ({ hash }: TxDetailsProps) => {
         </Alert>
       </Snackbar>{" "}
 
-    </Card>
+    </Card >
   );
 }