+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * Linked real-time data in Account Section
 * [#60] Changed footer logo & social links to Big Dipper
 * [#59] Unified layout on the desktop 
+* Added Signers in BlockDetails
 
 
 ## v0.1.0